Optional cursorword highlighting in insert mode
Cursorword highlighting may not be desired when in insert mode. When
cursorword highlighting uses fg and bg options particularly, the
highlighting of the current word feels distracting. For example if the
user is typing the word "return", highlighting triggers for "ret",
"retu", "retur" and finally "return". Attaching example screenshot to
ticket.